pyproject.toml
[tool.poetry]
name = "oulad-case"
version = "0.1.0"
description = ""
authors = ["Manuel Garcia <31109774+Mandroide@users.noreply.github.com>"]
packages = [{ include = "oulad_etl", from = "src" }]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2"
numpy = "^2.0"
pyarrow = "^20.0"
sqlalchemy = {extras = ["mysql"], version = "^2.0"}  # or use mysql-connector-python
mysql-connector-python = "^9.3.0"
pydantic-settings = "^2.9.1"
python-dotenv = "^1.0"
pyyaml = "^6.0.2"
rich = "^14.0.0"
requests = "^2.32.4"
tabulate = "^0.9.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "*"
jupyter = "*"
pytest = "^8.2"
pytest-cov = "^5.0"
black = "^24.4"
ruff = "^0.4"
mypy = "^1.10"
pre-commit = "^3.7"
detect-secrets = "^1.5"
types-requests = "^2.32.4.20250611"
types-pyyaml = "^6.0.12.20250516"
types-tabulate = "^0.9.0.20241207"

[tool.poetry.scripts]
etl = "oulad_etl.cli:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
